Move entityType into seperate file
